Skillmons: Lower secondary effect change depending on accuracy
To make the code slightly more readable, `init` was removed, and the
base power modifications are now performed during damage calculation
instead of during module initiailization. The alternatives would be
to either perform complex chance modifications during initialization,
but the code for that is quite complex, so repeating it is probably
not a good idea.