Add very rare TMTRAINER event for +xfix.
Before each hit, random number from 0 to 776 is generated. If it turns
out to be 42, TMTRAINER is enabled as a result of Super Glitch.

I don't expect many people to discover this (although somebody probably
will). Consider this to be a rarely occuring easter egg.

The behaviour of TMTRAINER is not like in Gen1, obviously.